requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
slack-bolt==1.18.0
slack-sdk==3.26.1
python-dotenv==1.0.0
openai==1.3.5
PyMuPDF==1.23.7
tinydb==4.8.0
python-multipart==0.0.6
requests==2.31.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
pytest-cov==4.1.0